requirements.txt
streamlit
pandas
langchain
langchain-openai
pydantic
tqdm
sqlalchemy
psycopg2-binary
faiss-cpu
sentence-transformers
python-dotenv
plotly
